// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/gpsLogUpsertRequest.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface GpsLogUpsertRequest {
  /**
   * @minimum -90
   * @maximum 90
   */
  lat: number;
  /**
   * @minimum -180
   * @maximum 180
   */
  lng: number;
  /**
   * @minimum 0
   * @maximum 250
   */
  speedKmh?: number;
  /**
   * @minimum 0
   * @maximum 360
   */
  bearing?: number;
}