// frontend/rodia-monorepo/apps/mobile/src/shared/api/generated/schemas/uploadDriverPhotoType.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type UploadDriverPhotoType = typeof UploadDriverPhotoType[keyof typeof UploadDriverPhotoType];


export const UploadDriverPhotoType = {
  PICKUP: 'PICKUP',
  DELIVERY: 'DELIVERY',
} as const;